code/requirements.txt
pyserial
numpy
pandas
scikit-learn
matplotlib
pywavelets
xgboost
jupyter
